+++ src/util/CustomPlayer.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <QStringView>
+
+namespace chatterino {
+
+void openInCustomPlayer(QStringView channel);
+
+}  // namespace chatterino