+++ src/widgets/settingspages/NotificationPage.cpp
@@ -127,7 +127,8 @@ QComboBox *NotificationPage::createToastReactionComboBox()
 {
     QComboBox *toastReactionOptions = new QComboBox();
 
-    for (int i = 0; i <= static_cast<int>(ToastReaction::DontOpen); i++)
+    for (int i = 0; i <= static_cast<int>(ToastReaction::OpenInCustomPlayer);
+         i++)
     {
         toastReactionOptions->insertItem(
             i, Toasts::findStringFromReaction(static_cast<ToastReaction>(i)));