+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Minor: When (re-)connecting, visible channels are now joined first. (#5850)
 - Minor: Added the ability to filter on messages by the author's user ID (example: `author.user_id == "22484632"`). (#5862)
 - Minor: Improved error messaging of the `/clip` command. (#5879)
+- Minor: Clicking on a live notification can now open the channel in a custom player. (#5880)
 - Minor: Added searchable hotkeys to global search and make keybinds searchable in the Hotkeys settings. (#5884)
 - Minor: Added Linux support for Live Notifications toasts. (#5881, #5971, #5976)
 - Minor: Messages can now be deleted from the context menu in a channel. (#5956)